+++ components/form-builder/elements/Paragraph.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import styled from "styled-components";
+
+const TextHint = styled.div`
+  margin: 20px 5px;
+  color: rgba(0, 0, 0, 0.38);
+`;
+
+export const Paragraph = () => {
+  return <TextHint>Long Answer text</TextHint>;
+};