+++ components/form-builder/icons/ParagraphIcon.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+export const ParagraphIcon = () => (
+  <svg width={16} height={14} fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M10 12H0v2h10v-2Zm6-8H0v2h16V4ZM0 10h16V8H0v2ZM0 0v2h16V0H0Z" fill="#5F6368" />
+  </svg>
+);