+++ components/form-builder/icons/RadioEmptyIcon.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+export const RadioEmptyIcon = () => (
+  <svg width={20} height={20} fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M10 0C4.48 0 0 4.48 0 10s4.48 10 10 10 10-4.48 10-10S15.52 0 10 0Zm0 18c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8Z"
+      fill="#5F6368"
+    />
+  </svg>
+);