+++ components/form-builder/icons/SelectMenuIcon.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+export const SelectMenuIcon = () => (
+  <svg width={20} height={20} fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M10 0C4.48 0 0 4.48 0 10s4.48 10 10 10 10-4.48 10-10S15.52 0 10 0Zm0 12L6 8h8l-4 4Z"
+      fill="#5F6368"
+    />
+  </svg>
+);