+++ components/form-builder/panel/index.ts
@@ -1,0 +1,2 @@
+export { ElementPanel } from "./ElementPanel";
+export { Button } from "./Button";